+++ src/main/java/ua/softserveinc/tc/entity/Room.java
@@ -53,7 +53,7 @@ public class Room {
             joinColumns = @JoinColumn(name = RoomConst.ID_ROOM))
     @MapKeyColumn(name = RoomConst.HOUR)
     @Column(name = RoomConst.PRICE)
-    private Map<Integer, Long> prices = new HashMap<>();
+    private Map<Integer, Double> prices = new HashMap<>();
 
     public Long getId() {
         return id;
@@ -119,11 +119,11 @@ public class Room {
         this.events = events;
     }
 
-    public Map<Integer, Long> getPrices() {
+    public Map<Integer, Double> getPrices() {
         return prices;
     }
 
-    public void setPrices(Map<Integer, Long> prices) {
+    public void setPrices(Map<Integer, Double> prices) {
         this.prices = prices;
     }
